Added an announcement of CTCSS (PL) at manual identification.
git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@454 f8843c6b-282e-0410-804b-c0d5895947d2

+++ src/svxlink/svxlink/MsgHandler.h
@@ -13,7 +13,7 @@ class MsgHandler : public SigC::Object
     
     void playFile(const std::string& path);
     void playMsg(const std::string& context, const std::string& msg);
-    void playNumber(int number);
+    void playNumber(float number);
     void spellWord(const std::string& word);
     void playSilence(int length);
     void writeBufferFull(bool is_full);
